Thermal subsystem second(?) draft
This version incorporates the working control system with 3 modes. On, off and maintain (the maintain could be removed).
It also uses the temperature to find the thermal exchange with the environment as a power loss.

A mask was created so you can edit many of the variables in the subsystem.

- NOT COVERED: assumptions are not yet covered and described there. There is no new media intake and only heating up old media. The "maintain" mode seem to be rather pointless as it only matters in a long period of time (if the heat coefficient is low).