Fluid system #3
In this draft the thermal and fluid system were attempted to be incorporated. It works pretty consistently with retentostats, but with chemostats some work may be needed.

Another change was the addition of both possibilities to be worked with in the fluid system. Now, the user can change the control between chemo and retentostats, although it is needed to do it manually. There are instructions on top of the control system of the fluid subsystem.

- THIS STILL DOES NOT USE ANY OD MEASUREMENT INPUTS AND THE RETENTOSTAT ESSENTIALLY ONLY INPUTS AND OUTPUTS THE SAME AMOUNT OF LIQUID/S AFTER REACHING A CERTAIN "OD CONDITION" SET BY A STEP SOURCE.