System changes
In this update there were changes to the thermal subsystem and liquid subsystem plus the creation of the integrated system file.

The changes for the thermal system were small, some parameters changed and the heat loss now happens between the outside and the wall of the vessel.
The fluid system changes were to fix the integration with the thermal system and a new control system for it.

The integrated system has a agitation system + all the other modifications of the two beforementioned subsystems. However, the agitation subsystem still does not work well.

Next steps will to begin to work on modelling the growth of the cells and the OD sensor. The other subsystems should remain unchanged, by my part, until the final control system (or a draft of it) is reached.