OD sensor integrated
Now the OD sensor has sucessfully(ish) been integrated. There is a description of the equations used for the growth based on the Monod's equation. There are a few things still to be seen.
What is the OD sensor actually measuring, units and a test with real values.
The in and out flow should both be positive!! That is why the absolute value is taken inside the growth subsystem.
The agitation system is still not working/integrated and some tidying needs to be done in terms of constants to unite all different relevant values.
The aeration subsystem still needs to be prepared and will be the next target towards integration.