Aeration system integrated
In this update the aeration system was created and integrated into the integrated model. It seems to be working so far.

The main assumption was made in how temperature changes the saturation of oxygen.  Since it involves a small range of temperatures (between 15 - 40 degrees) the change in saturation of oxygen may be regarded as a linear proportion. This can graphically be seen in the Ox_conc_linear_regression.m file.

The aeration system was integrated. There is no thermal insulation on the material and there is a control system for a second pump but it is not incorporated into the model.

Next steps will be to figure out the agitation system.