First compete system draft
In this update you can see the first draft of the complete integrated system of the bioreactor. The aeration system was also included as a separate subsystem file.

Next final steps will be to complete the  insulation aspect of the thermal system as it is not fully working. Also, add an "electrical power consumed" part that unites the whole system power requirements.
Also, an extra requirement is the "make it pretty" optional part. This should be done to clean everything up. For this part I am envisioning deleting all subsystem masks and creating a whole subsystem in the file that contains all variables that correlate to more than one subsystem.